backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic[email]==2.10.3
pydantic-settings==2.7.0
email-validator==2.2.0
python-dotenv==1.0.1
feedparser==6.0.11
httpx==0.28.1
beautifulsoup4==4.12.3
google-generativeai==0.8.3
scikit-learn==1.6.0
numpy==2.2.0
apscheduler==3.11.0
